+++ tests/bottleneck.rs
@@ -3,7 +3,7 @@
 // Licensed under the Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International
 // https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode
 
-use hashbrown::HashSet;
+use hashbrown::{HashMap, HashSet};
 use metropolis::agent::Agent;
 use metropolis::learning::{ExponentialLearningModel, LearningModel};
 use metropolis::mode::trip::{DepartureTimeModel, Leg, LegType, RoadLeg, TravelingMode};
@@ -28,7 +28,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
     let mut agents = Vec::with_capacity(departure_times.len());
     for (i, dt) in departure_times.into_iter().enumerate() {
         let leg = Leg::new(
-            LegType::Road(RoadLeg::new(node_index(0), node_index(2), vehicle_index(0))),
+            LegType::Road(RoadLeg::new(0, 2, vehicle_index(0))),
             Time::default(),
             TravelUtility::default(),
             ScheduleUtility::None,
@@ -76,6 +76,10 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
             ),
         ),
     ]);
+    let node_map: HashMap<_, _> = (0..=2)
+        .into_iter()
+        .map(|i| (i as u64, node_index(i)))
+        .collect();
     let vehicle = Vehicle::new(
         Length(1.0),
         PCE(1.0),
@@ -83,7 +87,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
         HashSet::new(),
         HashSet::new(),
     );
-    let road_network = RoadNetwork::new(graph, vec![vehicle]);
+    let road_network = RoadNetwork::new(graph, node_map, vec![vehicle]);
     let network = Network::new(Some(road_network));
 
     let parameters = Parameters {